linked cpu with interface. moved Makefile to source folder
