+++ cpu/CMakeLists.txt
@@ -1,6 +1,8 @@
 cmake_minimum_required(VERSION 3.1)
 set (CMAKE_CXX_STANDARD 11)
+set (CMAKE_SHARED_LINKER_FLAGS "-Wl,--as-needed,-Werror")
+set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -O0 -ggdb -Werror -Wall")
 project( cpu )
 add_executable( app application_opt.cpp )
 add_executable( verify verify.cpp sha256.cpp )
-add_executable( old application.cpp sha256.cpp )
+add_executable( old application.cpp sha256.cpp  ../interface/double_buffer.cpp)
